+++ cisco_acl/base_ace.py
@@ -54,4 +54,4 @@ class BaseAce(Base):
 
     @sequence.deleter
     def sequence(self) -> None:
-        del self.sequence.number
+        self._sequence = Sequence()