+++ tests_/test__package.py
@@ -8,6 +8,7 @@ from datetime import datetime
 # noinspection PyProtectedMember
 from cisco_acl import __title__
 from setup import PACKAGE, ROOT, README
+
 CHANGELOG = "CHANGELOG.rst"
 
 IMPORTS = [